+++ types.go
@@ -1,5 +1,7 @@
 package main
 
+import "sync"
+
 type Player struct {
 	Name    string
 	NameRaw []int
@@ -44,4 +46,27 @@ type Server struct {
 	QTV           []QTV
 
 	keepaliveCount int
+}
+
+type MutexStore struct {
+	sync.RWMutex
+	data []byte
+}
+
+func (store *MutexStore) Write(data []byte) {
+	store.Lock()
+	store.data = data
+	store.Unlock()
+}
+
+func (store *MutexStore) Read() []byte {
+	store.RLock()
+	data := store.data
+	store.RUnlock()
+	return data
+}
+
+func newMutexStore() *MutexStore {
+	store := MutexStore{data: make([]byte, 0)}
+	return &store
 }
